Phase 1: extract features/file-loading.js (file + URL loading)
handleFile/handleFileSelect/handleUrl + the URL-error helpers and VALID_EXTENSIONS
move out (showUrlError/clearUrlError are now module-private). Imports
normalizeMarkdownUrl + handleRepoUrl; createTab supplied via configureFileLoading
(DI var named openTab to avoid the eval-harness global collision). The drag
handlers stay in main and call the imported handleFile. main.js ~1,639 -> ~1,544.
Build + lint clean; 297 tests pass.
